Create README.md
This is my first python project. In this project i made a snake,water and gun game. 
The game is between the user and computer.
in this game snake is more powerful than water and water is more powerful than gun and gun is more powerful than snake..
So, if user choose snake and computer choose gun then computer is the winner..
Take a look of the game and tell me how much you like the game.